Add query validator

+++ src/Database/Database.php
@@ -14,7 +14,7 @@ class Database
     // Simple Types
     const VAR_STRING = 'string';
     const VAR_INTEGER = 'integer';
-    const VAR_FLOAT = 'float';
+    const VAR_FLOAT = 'double';
     const VAR_BOOLEAN = 'boolean';
     
     // Relationships Types
